[SPARK-9650][SQL] Fix quoting behavior on interpolated column names
